Moved code files from the previous architecture into "old" directory for both main code and tests
