+++ src/unifair/old/workflows/encode.py
@@ -1,8 +1,8 @@
 import os
 import sys
 
-from unifair.core.workflow import Workflow
-from unifair.steps.imports.encode import ImportEncodeMetadataFromApi
+from unifair.old.core.workflow import Workflow
+from unifair.old.steps.imports.encode import ImportEncodeMetadataFromApi
 
 
 def create_encode_workflow():
