+++ src/unifair/old/workflows/gdc.py
@@ -1,8 +1,8 @@
 import os
 import sys
 
-from unifair.core.workflow import Workflow
-from unifair.steps.imports.gdc import ImportGDCMetadataFromApi
+from unifair.old.core.workflow import Workflow
+from unifair.old.steps.imports.gdc import ImportGDCMetadataFromApi
 
 
 def create_gdc_workflow():
